classplayer.php renamed to classPlayer.php due to the way __autoloader() is written. the substring that reflects the name of the class has to match the class name exactly, including case-sensitivity. Also added getters to the Player class, and edited some syntax errors in the constructor.
player0.php I updated my code to match Sarah's latest changes, but I then added some functionality where $_SESSION data is pulled to load the user's username on the page. At a later date, the same will be done with stat info, but that is lower on the priority list.

validatelogin.php had to undergo a good handful of changes so that it could store a Player object as a $_SESSION variable. This will be useful in forming connections to different player for a game.