FEN_position.parsePiecePlacement has been written. (UNTESTED)
I re-defined the parse functions to be helper functions for FEN_position.formatCheckFEN, w/ formatCheckFEN updating the values of the private variables. Thus formatCheckFEN acts as a constructor for FEN_position.

I also am strongly considering renaming this file to fen and to also rename this branch fenParsingInJS to more accurately represent the work I'm doing.